requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
jinja2==3.1.4
pdfminer.six==20231228
PyPDF2==3.0.1
transformers==4.45.2
torch>=2.1.0
pydantic-settings==2.6.1
python-multipart==0.0.9
httpx==0.27.2
openai==1.51.2
